style: fix linting issues
