+++ packages/openid4vc-client/package.json
@@ -25,13 +25,12 @@
   },
   "dependencies": {
     "@aries-framework/core": "0.3.2",
-    "@sphereon/openid4vci-client": "^0.3.4",
+    "@sphereon/openid4vci-client": "^0.3.6",
     "class-transformer": "0.5.1",
     "class-validator": "0.13.1",
     "rxjs": "^7.2.0"
   },
-  "peerDependencies": {
-  },
+  "peerDependencies": {},
   "devDependencies": {
     "@aries-framework/node": "0.3.2",
     "reflect-metadata": "^0.1.13",
