+++ packages/openid4vc-client/src/OpenId4VcClientService.ts
@@ -1,7 +1,7 @@
 import { injectable, W3cCredentialService } from '@aries-framework/core'
 
 @injectable()
-export class OidcClientService {
+export class OpenId4VcClientService {
   private w3cCredentialService: W3cCredentialService
 
   public constructor(w3cCredentialService: W3cCredentialService) {
