+++ packages/openid4vc-client/src/index.ts
@@ -1,3 +1,3 @@
-export * from './OidcClientModule'
-export * from './OidcClientApi'
-export * from './OidcClientService'
+export * from './OpenId4VcClientModule'
+export * from './OpenId4VcClientApi'
+export * from './OpenId4VcClientService'
